Further refactoring within FEM, visualize, material and concepts
